[Setting] Clarity 추가
