+++ app/build.gradle.kts
@@ -38,7 +38,6 @@ android {
             )
             buildConfigField("String", "KAKAO_APP_KEY", getLocalProperty("KAKAO_APP_KEY"))
             resValue("string", "kakao_oauth_key", getLocalProperty("KAKAO_OAUTH_KEY"))
-            resValue("string", "clarity_key", getLocalProperty("CLARITY_KEY"))
         }
         debug {
             isMinifyEnabled = false
@@ -51,7 +50,6 @@ android {
             )
             buildConfigField("String", "KAKAO_APP_KEY", getLocalProperty("KAKAO_APP_KEY"))
             resValue("string", "kakao_oauth_key", getLocalProperty("KAKAO_OAUTH_KEY"))
-            resValue("string", "clarity_key", getLocalProperty("CLARITY_KEY"))
         }
     }
