+++ app/src/main/kotlin/ac/dnd/mour/android/MourApplication.kt
@@ -20,8 +20,6 @@ import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.crashlytics.ktx.crashlytics
 import com.google.firebase.ktx.Firebase
 import com.kakao.sdk.common.KakaoSdk
-import com.microsoft.clarity.Clarity
-import com.microsoft.clarity.ClarityConfig
 import dagger.hilt.android.HiltAndroidApp
 import io.sentry.Sentry
 import javax.inject.Inject
@@ -50,7 +48,6 @@ open class MourApplication : Application() {
         }
 
         initializeFirebase()
-        initializeClarity()
         observeRefreshTokenValidation()
     }
 
@@ -81,11 +78,6 @@ open class MourApplication : Application() {
 
     private fun initializeFirebase() {
         Firebase.analytics
-    }
-
-    private fun initializeClarity() {
-        val config = ClarityConfig(getString(R.string.clarity_key))
-        Clarity.initialize(applicationContext, config)
     }
 
     private fun observeRefreshTokenValidation() {
